rename files: test_rpki_parser, move helper module to util folder
